OOP PHP+Html example
In this example, I present the basic idea of having a new instance of a class, in which you define text through properties (could be links to other resources as well). The CSS could be included from a remote location or via the standard PHP functions.

There could be a lot of different ways to get the following result, including magic methods and using a single function (to make your life easier).