Singleton
Let's assume you want to have a single instance of a class (lets suggest you have a DB and you don't want a ton of instances flying all around, just because).
Singleton lets you create a single instance of a class. 
Calling another instance of the same class will be a futile.